Revamp: Move helpers inside cmds
